delete

+++ git-webhook.php
@@ -27,7 +27,9 @@ if ($content['ref'] == 'refs/heads/master') {
 	file_put_contents(LOG_DIR . "git-webhook.log", "****写入日志****" . PHP_EOL, FILE_APPEND);
 	//PHP函数执行git命令
 	$bool = chdir(SRC_DIR);
-	
+	if ($bool === false) {
+		die("Could not chdir()");
+	}
 	$re = shell_exec("ls -al");
 	$res = shell_exec('git reset --hard origin/master && git clean -f && git pull 2>&1 && git checkout master');
 	file_put_contents(LOG_DIR . "git-content.log", ($bool ? 'T' : 'F') . PHP_EOL, FILE_APPEND);
